+++ R/methods_fixest.R
@@ -1,50 +1,42 @@
 # .fixest -----------------------
 
 #' @export
-standard_error.fixest <- function(model, ...) {
-  stats <- summary(model)
+standard_error.fixest <- function(model, vcov = NULL, vcov_args = NULL, ...) {
   params <- insight::get_parameters(model)
+
+  if (!is.null(vcov)) {
+    # we don't want to wrap this in a tryCatch because the `fixest` error is
+    # informative when `vcov` is wrong.
+    V <- insight::get_varcov(model, vcov = vcov, vcov_args = vcov_args)
+    SE <- sqrt(diag(V))
+
+  } else {
+    stats <- summary(model)
+    SE <- as.vector(stats$se)
+  }
 
   .data_frame(
     Parameter = params$Parameter,
-    SE = as.vector(stats$se)
+    SE = SE
   )
 }
-
-## TODO add ci_method later?
-
-#' @export
-p_value.fixest <- function(model, ...) {
-  stats <- summary(model)$coeftable
-  params <- insight::get_parameters(model)
-  stat_col <- which(colnames(stats) %in% c("Pr(>|t|)", "Pr(>|z|)"))
-
-  .data_frame(
-    Parameter = params$Parameter,
-    p = as.vector(stats[, stat_col])
-  )
-}
-
 
 
 #' @export
 degrees_of_freedom.fixest <- function(model, method = "wald", ...) {
+  # fixest degrees of freedom can be tricky. best to use the function by the
+  # package.
   if (is.null(method)) {
     method <- "wald"
   }
-  method <- match.arg(tolower(method), choices = c("analytical", "any", "fit", "wald", "residual", "normal"))
-
-  if (method %in% c("wald", "residual", "fit")) {
-    s <- summary(model)
-    vcov_scaled <- s$cov.scaled
-    if (is.null(vcov_scaled)) {
-      s$nobs - s$nparams
-    } else {
-      max(s$nobs - attr(vcov_scaled, "dof.K"), 1)
-    }
-  } else {
-    degrees_of_freedom.default(model, method = method, ...)
-  }
+  method <- match.arg(
+    tolower(method),
+    choices = c("wald", "residual"))
+  method <- switch(
+    method,
+    "wald" = "t",
+    "residual" = "resid")
+  fixest::degrees_freedom(model, type = method)
 }